TST: Move merge/join tests into sub-dir
