+++ pandas/tests/reshape/merge/data/asof.csv
