+++ pandas/tests/reshape/merge/data/trades2.csv
