+++ pandas/tests/reshape/merge/test_merge_index_as_string.py
