+++ setup.py
@@ -761,6 +761,7 @@ setup(name=DISTNAME,
                 'pandas.tests.io.formats',
                 'pandas.tests.groupby',
                 'pandas.tests.reshape',
+                'pandas.tests.reshape.merge',
                 'pandas.tests.series',
                 'pandas.tests.scalar',
                 'pandas.tests.tseries',
@@ -800,6 +801,7 @@ setup(name=DISTNAME,
                     'pandas.tests.io.formats': ['data/*.csv'],
                     'pandas.tests.io.msgpack': ['data/*.mp'],
                     'pandas.tests.reshape': ['data/*.csv'],
+                    'pandas.tests.reshape.merge': ['data/*.csv'],
                     'pandas.tests.tseries.offsets': ['data/*.pickle'],
                     'pandas.io.formats': ['templates/*.tpl']
                     },
